+++ src/kusto/azext_kusto/vendored_sdks/kusto/_configuration.py
@@ -47,9 +47,8 @@ class KustoManagementClientConfiguration(Configuration):
 
         self.credential = credential
         self.subscription_id = subscription_id
-        self.api_version = "2020-09-18"
-        self.credential_scopes = ['https://management.azure.com/.default']
-        self.credential_scopes.extend(kwargs.pop('credential_scopes', []))
+        self.api_version = "2021-01-01"
+        self.credential_scopes = kwargs.pop('credential_scopes', ['https://management.azure.com/.default'])
         kwargs.setdefault('sdk_moniker', 'kustomanagementclient/{}'.format(VERSION))
         self._configure(**kwargs)