+++ src/virtual-wan/azext_vwan/custom.py
@@ -114,7 +114,7 @@ def create_virtual_wan(cmd, resource_group_name, virtual_wan_name, tags=None, lo
         office365_local_breakout_category=office365_category,
         type=vwan_type
     )
-    return client.create_or_update(resource_group_name, virtual_wan_name, wan)
+    return client.begin_create_or_update(resource_group_name, virtual_wan_name, wan)
 
 
 def update_virtual_wan(instance, tags=None, security_provider_name=None, branch_to_branch_traffic=None,
@@ -146,8 +146,35 @@ def create_virtual_hub(cmd, resource_group_name, virtual_hub_name, address_prefi
         virtual_wan=SubResource(id=virtual_wan),
         sku=sku
     )
-    return sdk_no_wait(no_wait, client.create_or_update,
+    return sdk_no_wait(no_wait, client.begin_create_or_update,
                        resource_group_name, virtual_hub_name, hub)
+
+
+def get_effective_virtual_hub_routes(cmd, resource_group_name, virtual_hub_name,
+                                     virtual_wan_resource_type=None, resource_id=None, no_wait=False):
+    parameters = None
+    Resource, EffectiveRoutesParameters = cmd.get_models("Resource", 'EffectiveRoutesParameters')
+    if virtual_wan_resource_type is not None or resource_id is not None:
+        parameters = EffectiveRoutesParameters(
+            virtual_wan_resource_type=virtual_wan_resource_type,
+            resource_id=Resource(id=resource_id)
+        )
+
+    client = network_client_factory(cmd.cli_ctx).virtual_hubs
+
+    def raw(response, *_):
+        import json
+        response = response.http_response
+        return json.loads(response.body())
+
+    return sdk_no_wait(
+        no_wait,
+        client.begin_get_effective_virtual_hub_routes,
+        resource_group_name,
+        virtual_hub_name,
+        parameters,
+        cls=raw
+    )
 
 
 def update_virtual_hub(instance, cmd, address_prefix=None, virtual_wan=None, tags=None, sku=None):
@@ -225,7 +252,7 @@ def create_hub_vnet_connection(cmd, resource_group_name, virtual_hub_name, conne
     )
 
     client = network_client_factory(cmd.cli_ctx).hub_virtual_network_connections
-    return sdk_no_wait(no_wait, client.create_or_update, resource_group_name,
+    return sdk_no_wait(no_wait, client.begin_create_or_update, resource_group_name,
                        virtual_hub_name, connection_name, connection)
 
 
@@ -236,7 +263,7 @@ def add_hub_route(cmd, resource_group_name, virtual_hub_name, address_prefixes, 
     hub = client.get(resource_group_name, virtual_hub_name)
     route = VirtualHubRoute(address_prefixes=address_prefixes, next_hop_ip_address=next_hop_ip_address)
     hub.route_table.routes.append(route)
-    poller = sdk_no_wait(no_wait, client.create_or_update,
+    poller = sdk_no_wait(no_wait, client.begin_create_or_update,
                          resource_group_name, virtual_hub_name, hub)
     try:
         return poller.result().route_table.routes
@@ -316,7 +343,7 @@ def create_vhub_route_table(cmd, resource_group_name, virtual_hub_name, route_ta
 
         client = _v3_route_table_client(cmd.cli_ctx)
 
-    return sdk_no_wait(no_wait, client.create_or_update, resource_group_name,
+    return sdk_no_wait(no_wait, client.begin_create_or_update, resource_group_name,
                        virtual_hub_name, route_table_name, route_table)
 
 
@@ -334,28 +361,25 @@ def update_vhub_route_table(cmd, resource_group_name, virtual_hub_name, route_ta
         client = _v3_route_table_client(cmd.cli_ctx)
         route_table.labels = labels
 
-    return sdk_no_wait(no_wait, client.create_or_update, resource_group_name,
+    return sdk_no_wait(no_wait, client.begin_create_or_update, resource_group_name,
                        virtual_hub_name, route_table_name, route_table)
 
 
 def get_vhub_route_table(cmd, resource_group_name, virtual_hub_name, route_table_name):
-    from msrestazure.azure_exceptions import CloudError
+    from azure.core.exceptions import ResourceNotFoundError
     try:
         return _v3_route_table_client(cmd.cli_ctx)\
             .get(resource_group_name, virtual_hub_name, route_table_name)  # Get v3 route table first.
-    except CloudError as ex:
-        if ex.status_code == 404:
-            return _v2_route_table_client(cmd.cli_ctx)\
-                .get(resource_group_name, virtual_hub_name, route_table_name)  # Get v2 route table.
-
-        raise
+    except ResourceNotFoundError:
+        return _v2_route_table_client(cmd.cli_ctx)\
+            .get(resource_group_name, virtual_hub_name, route_table_name)  # Get v2 route table.
 
 
 def delete_vhub_route_table(cmd, resource_group_name, virtual_hub_name, route_table_name, no_wait=False):
     route_table = get_vhub_route_table(cmd, resource_group_name, virtual_hub_name, route_table_name)
     client = _route_table_client(cmd.cli_ctx, route_table)
 
-    return sdk_no_wait(no_wait, client.delete, resource_group_name, virtual_hub_name, route_table_name)
+    return sdk_no_wait(no_wait, client.begin_delete, resource_group_name, virtual_hub_name, route_table_name)
 
 
 def list_vhub_route_tables(cmd, resource_group_name, virtual_hub_name):
@@ -401,7 +425,7 @@ def add_hub_routetable_route(cmd, resource_group_name, virtual_hub_name, route_t
                          next_hop=next_hop)
         route_table.routes.append(route)
 
-    poller = sdk_no_wait(no_wait, client.create_or_update,
+    poller = sdk_no_wait(no_wait, client.begin_create_or_update,
                          resource_group_name, virtual_hub_name, route_table_name, route_table)
     try:
         return poller.result().routes
@@ -423,7 +447,7 @@ def remove_hub_routetable_route(cmd, resource_group_name, virtual_hub_name, rout
         raise CLIError('invalid index: {}. Index can range from 1 to {}'.format(index, len(route_table.routes)))
 
     client = _route_table_client(cmd.cli_ctx, route_table)
-    poller = sdk_no_wait(no_wait, client.create_or_update,
+    poller = sdk_no_wait(no_wait, client.begin_create_or_update,
                          resource_group_name, virtual_hub_name, route_table_name, route_table)
     try:
         return poller.result().routes
@@ -443,7 +467,7 @@ def _route_table_client(cli_ctx, route_table):
 
 
 def _v2_route_table_client(cli_ctx):
-    return network_client_factory(cli_ctx).virtual_hub_route_table_v2s
+    return network_client_factory(cli_ctx).virtual_hub_route_table_v2_s
 
 
 def _v3_route_table_client(cli_ctx):
@@ -456,7 +480,7 @@ def create_vpn_gateway(cmd, resource_group_name, gateway_name, virtual_hub,
                        location=None, tags=None, scale_unit=None,
                        asn=None, bgp_peering_address=None, peer_weight=None, no_wait=False):
     from msrestazure.azure_exceptions import CloudError
-    from .vendored_sdks.v2018_08_01.v2018_08_01.models.error_py3 import ErrorException
+    from azure.core.exceptions import AzureError as ErrorException
     client = network_client_factory(cmd.cli_ctx).vpn_gateways
     try:
         client.get(resource_group_name, gateway_name)
@@ -476,7 +500,7 @@ def create_vpn_gateway(cmd, resource_group_name, gateway_name, virtual_hub,
             'peerWeight': peer_weight
         }
     )
-    return sdk_no_wait(no_wait, client.create_or_update,
+    return sdk_no_wait(no_wait, client.begin_create_or_update,
                        resource_group_name, gateway_name, gateway)
 
 
@@ -548,7 +572,7 @@ def create_vpn_gateway_connection(cmd, resource_group_name, gateway_name, connec
         routing_configuration=routing_configuration
     )
 
-    return sdk_no_wait(no_wait, client.create_or_update, resource_group_name, gateway_name, connection_name, conn)
+    return sdk_no_wait(no_wait, client.begin_create_or_update, resource_group_name, gateway_name, connection_name, conn)
 
 
 def list_vpn_gateways(cmd, resource_group_name=None):
@@ -643,7 +667,7 @@ def create_vpn_site(cmd, resource_group_name, vpn_site_name, ip_address,
     )
     if not any([asn, bgp_peering_address, peer_weight]):
         site.bgp_properties = None
-    return sdk_no_wait(no_wait, client.create_or_update,
+    return sdk_no_wait(no_wait, client.begin_create_or_update,
                        resource_group_name, vpn_site_name, site)
 
 
@@ -722,7 +746,7 @@ def create_vpn_server_config(cmd, resource_group_name, vpn_server_configuration_
         )
     )
 
-    return sdk_no_wait(no_wait, client.create_or_update,
+    return sdk_no_wait(no_wait, client.begin_create_or_update,
                        resource_group_name, vpn_server_configuration_name, vpn_server_config)
 
 
@@ -782,7 +806,7 @@ def add_vpn_server_config_ipsec_policy(cmd, resource_group_name, vpn_server_conf
             pfs_group=pfs_group
         )
     )
-    poller = sdk_no_wait(no_wait, client.create_or_update,
+    poller = sdk_no_wait(no_wait, client.begin_create_or_update,
                          resource_group_name, vpn_server_configuration_name, vpn_server_config)
     if no_wait:
         return poller
@@ -804,7 +828,7 @@ def remove_vpn_server_config_ipsec_policy(cmd, resource_group_name, vpn_server_c
         vpn_server_config.vpn_client_ipsec_policies.pop(index)
     except IndexError:
         raise CLIError('invalid index: {}. Index can range from 0 to {}'.format(index, len(vpn_server_config.vpn_client_ipsec_policies) - 1))
-    poller = sdk_no_wait(no_wait, client.create_or_update,
+    poller = sdk_no_wait(no_wait, client.begin_create_or_update,
                          resource_group_name, vpn_server_configuration_name, vpn_server_config)
     if no_wait:
         return poller
@@ -816,7 +840,7 @@ def create_p2s_vpn_gateway(cmd, resource_group_name, gateway_name, virtual_hub,
                            scale_unit, location=None, tags=None, p2s_conn_config_name='P2SConnectionConfigDefault',
                            vpn_server_config=None, address_space=None, associated_route_table=None,
                            propagated_route_tables=None, labels=None, no_wait=False):
-    client = network_client_factory(cmd.cli_ctx).p2s_vpn_gateways
+    client = network_client_factory(cmd.cli_ctx).p2_svpn_gateways
     (P2SVpnGateway,
      SubResource,
      P2SConnectionConfiguration,
@@ -843,7 +867,7 @@ def create_p2s_vpn_gateway(cmd, resource_group_name, gateway_name, virtual_hub,
         virtual_hub=SubResource(id=virtual_hub) if virtual_hub else None,
         vpn_gateway_scale_unit=scale_unit,
         vpn_server_configuration=SubResource(id=vpn_server_config) if vpn_server_config else None,
-        p2_sconnection_configurations=[
+        p2_s_connection_configurations=[
             P2SConnectionConfiguration(
                 vpn_client_address_pool=AddressSpace(
                     address_prefixes=address_space
@@ -854,7 +878,7 @@ def create_p2s_vpn_gateway(cmd, resource_group_name, gateway_name, virtual_hub,
         ]
     )
 
-    return sdk_no_wait(no_wait, client.create_or_update, resource_group_name, gateway_name, gateway)
+    return sdk_no_wait(no_wait, client.begin_create_or_update, resource_group_name, gateway_name, gateway)
 
 
 def update_p2s_vpn_gateway(instance, cmd, tags=None, scale_unit=None,
@@ -865,7 +889,7 @@ def update_p2s_vpn_gateway(instance, cmd, tags=None, scale_unit=None,
         c.set_param('tags', tags, True)
         c.set_param('vpn_gateway_scale_unit', scale_unit, False)
         c.set_param('vpn_server_configuration', SubResource(id=vpn_server_config) if vpn_server_config else None, True)
-    p2_sconnection_configurations = getattr(instance, 'p2_sconnection_configurations')
+    p2_sconnection_configurations = getattr(instance, 'p2_s_connection_configurations')
     if p2_sconnection_configurations:
         with UpdateContext(p2_sconnection_configurations[0]) as c:
             c.set_param('vpn_client_address_pool.address_prefixes', address_space, False)
@@ -881,10 +905,21 @@ def update_p2s_vpn_gateway(instance, cmd, tags=None, scale_unit=None,
 
 
 def list_p2s_vpn_gateways(cmd, resource_group_name=None):
-    client = network_client_factory(cmd.cli_ctx).p2s_vpn_gateways
+    client = network_client_factory(cmd.cli_ctx).p2_svpn_gateways
     if resource_group_name:
         return client.list_by_resource_group(resource_group_name)
     return client.list()
+
+
+def generate_vpn_profile(cmd, resource_group_name, gateway_name, authentication_method=None):
+    client = network_client_factory(cmd.cli_ctx).p2_svpn_gateways
+    P2SVpnProfileParameters = cmd.get_models('P2SVpnProfileParameters')
+    parameters = P2SVpnProfileParameters(authentication_method=authentication_method)
+    return client.begin_generate_vpn_profile(
+        resource_group_name,
+        gateway_name,
+        parameters
+    )
 
 
 def _load_cert_file(file_path):